==> 3. Player Data Dataframe Consolidation.ipynb <==
import datetime
from datetime import date
import requests
import pprint
from bs4 import BeautifulSoup
import pandas as pd
import re
import pickle
from urllib.request import urlopen
import glob
import altair as alt


# Reading Dataframes 

df_final_1 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_1.csv')
df_final_2 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_2.csv')
df_final_3 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_3.csv')
df_final_4 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_4.csv')
df_final_5 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_5.csv')
df_final_6 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_6.csv')
df_final_7 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_7.csv')
df_final_8 = pd.read_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_8.csv')


print('df_final_1 shape is:', df_final_1.shape)
print('df_final_2 shape is:', df_final_2.shape)
print('df_final_3 shape is:', df_final_3.shape)
print('df_final_4 shape is:', df_final_4.shape)
print('df_final_5 shape is:', df_final_5.shape)
print('df_final_6 shape is:', df_final_6.shape)
print('df_final_7 shape is:', df_final_7.shape)
print('df_final_8 shape is:', df_final_8.shape)

# Concatenating df_final data frames

df_final_list = [df_final_1, df_final_2, df_final_3, df_final_4, df_final_5, df_final_6, df_final_7, df_final_8]
consolidated_df_final = pd.concat(df_final_list, axis=0, ignore_index=True)

# Cleaning NaN's from df_final

pd.options.display.max_columns = 50 # to set maximum diplay width

print(consolidated_df_final.shape)
print(len(consolidated_df_final['name'].unique()))

consolidated_df_final.dropna(axis = 0, subset=['Date'], inplace = True)
print(consolidated_df_final.shape)
print(len(consolidated_df_final['name'].unique()))

# testing Cristiano Ronaldo

consolidated_df_final

# Dropping unwanted columns from df_final

# consolidated_df_final.drop(columns = ['Match Report'], inplace = True)

# Converting date columns to datetime

consolidated_df_final['Date'] = pd.to_datetime(consolidated_df_final['Date'])

# Saving Consolidated Dataframe

pd.DataFrame(consolidated_df_final).to_csv('G:/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/consolidated_df_final.csv', index=False)
