==> 9. Injury Duration Prediction.ipynb <==
import datetime
from datetime import date
from datetime import datetime
import requests
import pprint
from bs4 import BeautifulSoup
import pandas as pd
import re
import pickle
from urllib.request import urlopen
import glob
# import altair as alt
import time
import itertools
import numpy as np
import operator
from datetime import timedelta
from ipywidgets import interact, IntSlider
from IPython.display import display
import sklearn
import numpy as np
from sklearn.linear_model import LogisticRegression
# import seaborn as sns
# from matplotlib import pyplot as plt
# %matplotlib inline
from sklearn.metrics import accuracy_score, confusion_matrix,roc_curve, roc_auc_score, precision_score, recall_score, precision_recall_curve
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score, RepeatedStratifiedKFold, StratifiedKFold
from sklearn.datasets import make_hastie_10_2
from sklearn.ensemble import GradientBoostingClassifier
from imblearn.pipeline import make_pipeline as make_pipeline_with_sampler
from imblearn.under_sampling import RandomUnderSampler

# from pycaret.classification import * 
from pycaret.regression import * 

# pd.options.display.max_columns = 5000
# pd.options.display.max_rows = 5000

injured_pred = 'injury_duration'

# simple_features = ['Height', 'Weight', 'age','cum_injury_total', 'weeks_since_last_injury', 'Min_cum','Serie A_cum',
#  'Premier League_cum', 'La Liga_cum', 'Ligue 1_cum', 'Bundesliga_cum', 'Champions Lg_cum', 'Europa Lg_cum', 'FIFA World Cup_cum', 'UEFA Nations League_cum', 'UEFA Euro_cum',
#  'Copa América_cum', 'Away_cum', 'Home_cum', 'Neutral_cum']

# Load Dataset

# Renzo's Path
# dataset = pd.read_csv('/Volumes/GoogleDrive/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/complete_final_df_5.csv')
#
# Sravan's Path
# dataset = pd.read_csv('/Users/srav88/My Drive/SIADS 697 - Capstone/Dataframes/dataset_for_model_final.csv')

# Sravan's Parquet Path
dataset = pd.read_parquet('dataframes_blog/dataset_for_model_final.parquet') 

injury_duration_df = dataset[['FBRefID', 'cum_injury_total', 'Injury','injury_count']].groupby(['FBRefID', 'cum_injury_total', 'Injury']).sum().reset_index()
injury_duration_df = injury_duration_df[injury_duration_df['Injury'] != '0'].rename(columns={'injury_count':'injury_duration'})
new_player_df_1 = pd.merge(dataset, injury_duration_df, left_on=['FBRefID', 'cum_injury_total', 'Injury'], right_on=['FBRefID', 'cum_injury_total', 'Injury'], how='left')
new_player_df_1['injury_duration'] = new_player_df_1['injury_duration'].fillna(0)

dataset_injury = new_player_df_1[new_player_df_1['injury_duration']!=0]

dataset_injury = dataset_injury.drop_duplicates(['Injury', 'injury_week', 'injury_year'])
dataset_injury = dataset_injury[dataset_injury['Injury']!='Unknown Injury'].reset_index(drop=True)

feature_list = list(dict(dataset_injury['Injury'].value_counts()).keys())

for col in feature_list:
    dataset_injury.loc[dataset_injury['Injury'] == col, col] = 1
    dataset_injury.loc[dataset_injury['Injury'] != col, col] = 0

extended_features = ['Height', 'Weight', 'cum_week', 'defender', 'attacker', 'midfielder', 'goalkeeper', 'right_foot', 'cum_injury_total',
 'weeks_since_last_injury', 'Min_cum', 'Gls_cum', 'Ast_cum', 'PK_cum', 'PKatt_cum', 'Sh_cum', 'SoT_cum', 'CrdY_cum',
 'CrdR_cum', 'Touches_cum', 'Press_cum', 'Tkl_cum', 'Int_cum', 'Blocks_cum', 'xG_cum', 'npxG_cum', 'xA_cum',
 'SCA_cum', 'GCA_cum', 'Cmp_cum', 'Att_cum', 'Prog_cum', 'Carries_cum', 'Prog.1_cum', 'Succ_cum', 'Att.1_cum',
 'Fls_cum', 'Fld_cum', 'Off_cum', 'Crs_cum', 'TklW_cum', 'OG_cum', 'PKwon_cum', 'PKcon_cum', 'Serie A_cum', 'Premier League_cum',
 'La Liga_cum', 'Ligue 1_cum', 'Bundesliga_cum', 'Champions Lg_cum', 'Europa Lg_cum', 'FIFA World Cup_cum',
 'UEFA Nations League_cum', 'UEFA Euro_cum', 'Copa América_cum', 'Away_cum', 'Home_cum', 'Neutral_cum',
                     'Injury_Recurrence_cum'] + list(dict(dataset_injury['Injury'].value_counts()).keys())[:30]

injury_recurrence = pd.DataFrame(dataset_injury.groupby(['FBRefID','Injury', 'injury_week', 'injury_year'])['Injury'].count()).rename(columns={'Injury':'Injury_Recurrence'}).reset_index().sort_values(by=['FBRefID','injury_year', 'injury_week'])
injury_recurrence['Injury_Recurrence_cum'] = injury_recurrence.groupby(['FBRefID','Injury'])['Injury_Recurrence'].cumsum()


injury_recurrence

dataset_injury_1 = pd.merge(dataset_injury, injury_recurrence, on=['FBRefID', 'Injury', 'injury_week', 'injury_year'], how='left')


exp_reg = setup(dataset_injury_1[extended_features + [injured_pred]], target = injured_pred, fold=10, feature_selection=True)


best_model = compare_models(turbo=False)

# df = pull()
# df.to_csv('results_1_year.csv', index=False)
df = pull().sort_values(by=['F1'], ascending=False)

df

# Select best model
model = create_model(df.index[0], fold=5)
save_model(model, 'model_1_year')

tuned_model = tune_model(model, optimize = 'F1')
save_model(tuned_model, 'model_1_year_tuned')

# Loading the saved model
model_1_year_tuned = load_model('model_1_year_tuned')

model_1_year_tuned
