==> 13. FBREF Player Batch 80000-84959.ipynb <==
import datetime
from datetime import date
import requests
import pprint
from bs4 import BeautifulSoup
import pandas as pd
import re
import pickle
from urllib.request import urlopen


# Reading CSV File 

match_logs_list_urls = pd.read_csv('/Volumes/GoogleDrive/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/CSV Files/match_logs_list_urls.csv')
# match_logs_list_urls = pd.read_csv('/Users/vruiz.CDS/Downloads/Capstone Files/match_logs_list_urls.csv')


# This one works

match_logs_list_urls = list(match_logs_list_urls['0'])
match_logs_list_urls = ['https://fbref.com' + i for i in match_logs_list_urls]


# Pull all match_log_lists_x tables. We will convert each list individually WORK IN PROCESS

def create_match_logs_tables(match_logs_list_urls_x):

    df_30_columns = pd.DataFrame([])
    df_39_columns = pd.DataFrame([])

    count = 0

    for player in match_logs_list_urls_x:
        try: # this may fix "HTTP Error 404: Not Found"
            # urlopen(player)

            new_table = pd.read_html(player)[0]
            new_table.columns = new_table.columns.droplevel()
            new_table['name'] = player.split('/')[-1].replace("-Match-Logs", "")
            
            if new_table.shape[1] == 30:
                new_table['FBRefID'] = player[(player.find("players/") + len("players/")):(player.find("/matchlogs"))]
                df_30_columns = df_30_columns.append(new_table, ignore_index=True)
                count += 1
                
                
            if new_table.shape[1] == 39:
                new_table['FBRefID'] = player[(player.find("players/") + len("players/")):(player.find("/matchlogs"))]
                df_39_columns = df_39_columns.append(new_table, ignore_index=True)
                count += 1

            sys.stdout.write("\r{0} percent player urls have just scraped!".format(count / len(match_logs_list_urls_x)*100))
            sys.stdout.flush()
                

        except:
            pass
    

    return df_30_columns, df_39_columns



len(match_logs_list_urls)

# Creating different length data frames

df_30_columns_1, df_39_columns_1 = create_match_logs_tables(match_logs_list_urls[80000:-1])

# export to google drive as df_30_columns_12 and df_columns_12

pd.DataFrame(df_30_columns_1).to_csv('/Volumes/GoogleDrive/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_30_columns_12.csv', index = False)
pd.DataFrame(df_39_columns_1).to_csv('/Volumes/GoogleDrive/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_39_columns_12.csv', index = False)

# pd.DataFrame(df_30_columns_1).to_csv('/Users/vruiz.CDS/Downloads/Capstone Files/df_30_columns_12.csv', index = False)
# pd.DataFrame(df_39_columns_1).to_csv('/Users/vruiz.CDS/Downloads/Capstone Files/df_39_columns_12.csv', index = False)

#Combining Df_30_columns_1 and df_39_columns_1 to dataframe_12

cols = ['Date', 'Day', 'Comp', 'Round', 'Venue', 'Result', 'Squad', 'Opponent', 'Start', 'Pos', 'Min', 'Gls', 'Ast', 'PK', 'PKatt', 'Sh', 'SoT', 'CrdY',
       'CrdR', 'Match Report', 'Int', 'name', 'FBRefID']

df1 = df_39_columns_1
df2 = df_30_columns_1

df_final_12 = df1.merge(df2,how='outer', left_on=cols, right_on=cols)

print(df1.shape)
print(df2.shape)
# print(df_final.shape)

# export to google drive as dataframe_12

pd.DataFrame(df_final_12).to_csv('/Volumes/GoogleDrive/.shortcut-targets-by-id/1KUGn_35OjAoOP2puz6yG-2g_8LBxvDG_/SIADS 697 - Capstone/Dataframes/df_final_12.csv', index=False)

# pd.DataFrame(df_39_columns_1).to_csv('/Users/vruiz.CDS/Downloads/Capstone Files/df_final_12.csv', index = False)
